fix(audit): batch 5 standalone P1s (RB + DS + SEC)
Closes 5 P1 audit findings, each independent + easy:

- **RB-V1.40-1**: `TestMapArgs::depth` had no clap range bound — `cqs
  test-map foo --depth 18446744073709551615` flowed through to BFS
  arithmetic where `chain_limit = max_depth + 1` panicked in debug.
  Two-layer fix: (1) `value_parser = clap::value_parser!(u16).range(1..=50)`
  matching `TraceArgs::max_depth`. (2) Defensive `saturating_add(1)` at
  the BFS arithmetic for direct lib callers bypassing clap. Type cascade:
  `DEFAULT_DEPTH_TEST_MAP` and `TestMapArgs::depth` flip from `usize`
  to `u16`; CLI/daemon dispatchers cast to `usize` at the boundary.

- **RB-V1.40-3**: `classify_hits` ended with `*kinds.iter().next().expect(
  "non-empty single-kind set")`. Logically safe today, but `.expect`
  carries the same panic semantics as `.unwrap` and the project's
  "no unwrap outside tests" rule makes that a code smell. Replaced
  with `.unwrap_or(Kind::Other)` — same observable behavior on the
  happy path; future refactor that breaks the invariant routes to
  the catch-all variant rather than crashing.

- **DS-V1.40-2**: `cmd_telemetry_reset` was non-atomic — `fs::copy(
  current, archive); fs::write(current, reset_event)`. A kill between
  the two calls left either a duplicate archive or a truncated current
  with the reset event lost. Both modes silently corrupt the autopilot
  measurement-window contract per the [Reset Telemetry After Autopilot]
  memory rule. Fix: atomic `rename(current → archive)` (POSIX rename(2)
  / Windows MoveFileExW with REPLACE_EXISTING) + tempfile-staged reset
  event with `cqs::fs::atomic_replace`. Kill between rename and replace
  leaves no current file → next call's `if !current.exists()` short-
  circuit handles it cleanly.

- **DS-V1.40-4**: `evals/regenerate_v3_test.py` wrote the v3.v2 fixture
  via `Path.write_text` (open + write + close — not atomic). Ctrl+C /
  OOM mid-write left a truncated fixture, which `cqs eval` either
  failed to parse (loud) or partially loaded (silent — produced wrong
  recall numbers). Per [Paired Eval for Retunes] memory rule, fixture
  fidelity is the ground truth for retune ship/no-ship decisions.
  Fix: helper `_atomic_write_text` writes to `<path>.tmp` then
  `os.replace` (atomic on POSIX rename(2) / Windows). Used for both
  fixture and diff writes.

- **SEC-V1.40-2**: `redact_userinfo` searched for the FIRST `@` in
  the entire after-scheme portion of a URL. For `https://api.com/path/@anchor`
  (a URL with `@` in the path), it produced `https://anchor` —
  silently rewriting the host to a path fragment. Fix: confine the
  `@` search to the authority component (terminating at first `/`
  or `?`), per RFC 3986. Added 3 regression tests covering: path
  `@`, query `@`, and the mixed case (userinfo in authority + `@`
  in path). All pre-existing tests still pass; new shape is strictly
  more correct.

Verification:
- `cargo build --features gpu-index` clean
- `cargo clippy -- -D warnings` clean
- `cargo fmt --check` clean
- `cargo test --features gpu-index --lib 'kind::tests'` 14/14 pass
- `cargo test --features gpu-index --lib 'llm::tests::redact_userinfo'` 4/4 pass
- `python3 -c 'import ast; ast.parse(open(...))'` clean

+++ src/cli/batch/handlers/graph.rs
@@ -435,7 +435,7 @@ pub(in crate::cli::batch) fn dispatch_test_map(
     args: &TestMapArgs,
 ) -> Result<serde_json::Value> {
     let name = args.name.as_str();
-    let max_depth = args.depth;
+    let max_depth = args.depth as usize;
     let cross_project = args.cross_project;
     let _span = tracing::info_span!(
         "batch_test_map",
